[changed] Introducing ButtonInput
Encapsulating types "button", "reset" and "submit" into a single
component. Usage of these types with Input is now deprecated. Please use
ButtonInput instead.
